+++ src/interceptors.ts
@@ -4,13 +4,13 @@
 
 import { JaFetchRequestInit } from './service';
 
-type ReqOnFulfilled = (url: string, init: JaFetchRequestInit) => JaFetchRequestInit;
-type ResOnFulfilled = (data: any, requestInfo: { url: string; config: JaFetchRequestInit }, response: Response) => void;
-type OnRejected = (error: any) => Promise<any>;
-/** */
+type ReqOnFulfilled = (url?: string, init?: JaFetchRequestInit) => JaFetchRequestInit;
+type ResOnFulfilled = (data?: any, request?: { url: string; init: JaFetchRequestInit }, response?: Response) => void;
+type OnRejected = (error?: any) => Promise<any>;
 type Store<T, U> = { id: number; onFulfilled: T; onRejected: U }[];
 
 class Interceptor<T, U> {
+  /**保存use方法中，传入的拦截方法 */
   store: Store<T, U> = [];
   onFulfilled: T;
   onRejected: U;
@@ -20,7 +20,7 @@ class Interceptor<T, U> {
    * @param onRejected
    * @returns
    */
-  use(onFulfilled: T, onRejected: U): number {
+  use(onFulfilled: T, onRejected?: U): number {
     if (onFulfilled && typeof onFulfilled !== 'function') {
       throw new TypeError('interceptor.use(onFulfilled, onRejected), parameter onFulfilled is not a function');
     }
@@ -39,7 +39,7 @@ class Interceptor<T, U> {
    * if !id , remove all Interceptor
    * @param {number} id
    */
-  remove(id: number) {
+  remove(id?: number) {
     if (id) {
       this.store = this.store.filter(it => it.id === id);
     } else {
@@ -51,13 +51,22 @@ class Interceptor<T, U> {
 export default class Interceptors {
   request = new Interceptor<ReqOnFulfilled, OnRejected>();
   response = new Interceptor<ResOnFulfilled, OnRejected>();
+
+  create() {
+    return new Interceptors();
+  }
   /**
-   * 拦截器预设
+   * 拦截器预设,use 另一个拦截器
    * @param {{install(interceptors:Interceptors)}} obj
    */
-  use(obj: { install: (interceptors: Interceptors) => void }) {
-    const interceptors = new Interceptors();
-    obj.install(interceptors);
+  use(obj: Interceptors | { install: (interceptors: Interceptors) => void }) {
+    let interceptors;
+    if (obj instanceof Interceptors) {
+      interceptors = obj;
+    } else {
+      interceptors = new Interceptors();
+      obj.install(interceptors);
+    }
 
     interceptors.request.store.forEach(item => {
       this.request.use(item.onFulfilled, item.onRejected);
@@ -66,8 +75,5 @@ export default class Interceptors {
       this.response.use(item.onFulfilled, item.onRejected);
     });
     // release interceptors?
-  }
-  create() {
-    return new Interceptors();
   }
 }