+++ types/interceptors.d.ts
@@ -1,10 +1,10 @@
 import { JaFetchRequestInit } from './service';
-declare type ReqOnFulfilled = (url: string, init: JaFetchRequestInit) => JaFetchRequestInit;
-declare type ResOnFulfilled = (data: any, requestInfo: {
+declare type ReqOnFulfilled = (url?: string, init?: JaFetchRequestInit) => JaFetchRequestInit;
+declare type ResOnFulfilled = (data?: any, request?: {
     url: string;
-    config: JaFetchRequestInit;
-}, response: Response) => void;
-declare type OnRejected = (error: any) => Promise<any>;
+    init: JaFetchRequestInit;
+}, response?: Response) => void;
+declare type OnRejected = (error?: any) => Promise<any>;
 declare type Store<T, U> = {
     id: number;
     onFulfilled: T;
@@ -14,15 +14,15 @@ declare class Interceptor<T, U> {
     store: Store<T, U>;
     onFulfilled: T;
     onRejected: U;
-    use(onFulfilled: T, onRejected: U): number;
-    remove(id: number): void;
+    use(onFulfilled: T, onRejected?: U): number;
+    remove(id?: number): void;
 }
 export default class Interceptors {
     request: Interceptor<ReqOnFulfilled, OnRejected>;
     response: Interceptor<ResOnFulfilled, OnRejected>;
-    use(obj: {
+    create(): Interceptors;
+    use(obj: Interceptors | {
         install: (interceptors: Interceptors) => void;
     }): void;
-    create(): Interceptors;
 }
 export {};